refactor: add new package (statistics)
