+++ app/src/main/java/ac/sbmax002/eye_on/ui/statistics/StatisticsScreen.kt
@@ -383,3 +383,4 @@ fun StatisticsTopBar(title: String, onBackClick: () -> Unit) {
         colors = TopAppBarDefaults.topAppBarColors(containerColor = Color(0xFF0A0A0A))
     )
 }
+
